remove Bowser requirement
There was originally implementation for browser detection, however this was never really used for anything other than just opening an ACTUAL new window if the browser wasn't chrome. 
The dependency was Bowser, which no longer is included in the repository or any of the code, so it should not be referenced here.